requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
aiosqlite
aisqlite
alembic
annotated-types
anyio
asttokens
asyncpg
attrs
av
certifi
cffi
charset-normalizer
click
colorama
coloredlogs
comm
cryptography
ctranslate2
debugpy
decorator
deprecation
distro
dnspython
docutils
ecdsa
email_validator
enum34
evdev
executing
fastapi
faster-whisper
filelock
flatbuffers
frozenlist
fsspec
gotrue
greenlet
groq
h11
h2
halo
hpack
httpcore
httptools
httpx
huggingface-hub
humanfriendly
hyperframe
id
idna
IProgress
ipykernel
ipython
ipython_pygments_lexers
jaraco.classes
jaraco.context
jaraco.functools
jedi
jeepney
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jupyter_client
jupyter_core
jwt
keyboard
keyring
langchain
langchain-core
langchain-groq
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-checkpoint-postgres
langgraph-prebuilt
langgraph-sdk
langsmith
log-symbols
Mako
markdown-it-py
MarkupSafe
matplotlib-inline
mdurl
more-itertools
mpmath
msgpack
multidict
nest-asyncio
networkx
nh3
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
onnxruntime
openai
openwakeword
orjson
packaging
parso
passlib
pexpect
pillow
platformdirs
postgrest
prompt_toolkit
propcache
protobuf
psutil
psycopg
psycopg-binary
psycopg-pool
ptyprocess
pure_eval
pvporcupine
pyasn1
PyAudio
pycparser
pydantic
pydantic_core
Pygments
PyJWT
pynput
python-dateutil
python-dotenv
python-jose
python-multipart
python-xlib
PyYAML
pyzmq
readme_renderer
RealtimeSTT
redis
requests
requests-toolbelt
rfc3986
rich
rsa
scikit-learn
scipy
SecretStorage
setuptools
six
sniffio
soundfile
spinners
SQLAlchemy
stack-data
starlette
storage3
StrEnum
supabase
supafunc
sympy
tenacity
termcolor
threadpoolctl
tokenizers
torch
torchaudio
torchvision
tornado
tqdm
traitlets
triton
twine
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
wcwidth
webrtcvad-wheels
websocket-client
websockets
wheel
yarl
zstandard
